requirements.txt
streamlit
opencv-python==4.8.0.76
numpy==1.26.0
Pillow==10.0.0